pyproject.toml
[project]
name = "mybot"
version = "0.1.0"
description = "mybot"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "yt-dlp>=2025.01.01",  # 新增：视频下载核心
    "httpx>=0.27.0",       # 新增：网络请求库
]

[tool.nonebot]
project_config = { } # 关键：告诉 NoneBot 去哪里加载你的插件
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo"]

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
